+++ test/test_parser.py
@@ -216,6 +216,16 @@ def test_parser_from_isbn():
     assert entry.data['year'] == '2013'
 
 
+# regression test for https://gitlab.com/mrossinek/cobib/-/issues/53
+def test_parser_from_isbn_empty():
+    """Test parsning an empty ISBN."""
+    root = path.abspath(path.dirname(__file__))
+    CONFIG.set_config(Path(root + '/../cobib/docs/debug.ini'))
+    entries = parser.Entry.from_isbn('3860704443')
+    assert not entries
+    assert entries == {}
+
+
 def test_parser_from_arxiv():
     """Test parsing from arxiv."""
     root = path.abspath(path.dirname(__file__))